requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
Django==3.0
django-model-utils==4.0.0
djangorestframework==3.11.0
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
pytz==2019.3
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
